Change type for latitude and longitude

+++ app/src/main/java/org/d3ifcool/hystorms/network/WeatherRetrofit.kt
@@ -6,8 +6,8 @@ import retrofit2.http.Query
 interface WeatherRetrofit {
     @GET("weather")
     suspend fun get(
-        @Query("lat") latitude: Long,
-        @Query("lon") longitude: Long,
+        @Query("lat") latitude: Double,
+        @Query("lon") longitude: Double,
         @Query("appid") appId: String,
         @Query("lang") language: String
     ): WeatherEntity
